Passage du salare en int

+++ Barman.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 
 public class Barman extends Employe{
 
-    Barman(String nom, String prenom, float salaire, ArrayList<String> edt){
+    Barman(String nom, String prenom, int salaire, ArrayList<String> edt){
         super(nom,prenom,salaire,edt);
     }
